strona/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
certifi==2019.9.11
chardet==3.0.4
Click==7.0
distlib==0.3.0
Django==3.0.3
django-braces==1.14.0
django-crispy-forms==1.9.0
django-generic-aggregation==0.4.0
django-model-utils==4.0.0
django-simple-ratings==0.3.3
django-star-ratings==0.9.1
django-taggit==1.2.0
docutils==0.15.2
filelock==3.0.12
idna==2.8
importlib-metadata==1.5.0
Kivy==1.11.1
kivy-deps.angle==0.1.9
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.17
kivy-deps.sdl2==0.1.22
Kivy-Garden==0.1.4
kivy-garden.graph==0.4.dev0
numpy==1.17.3
Pillow==6.2.1
pygame==1.9.6
Pygments==2.4.2
pypiwin32==223
PyQt5==5.13.0
PyQt5-sip==12.7.0
pyqt5-tools==5.13.0.1.5
PyQtWebEngine==5.13.2
python-dotenv==0.10.3
pytz==2019.3
pywin32==225
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
swapper==1.1.2.post1
urllib3==1.25.3
virtualenv==20.0.5
wxPython==4.0.7.post2
zipp==3.0.0
